Add CJS compat plugin to exec
